Move: Change project file structure
