+++ WorkUp/WorkUp/View/CardDetail/ProgressRingView.swift
