+++ WorkUp/WorkUp/View/CardDetail/YawMotionCardView/CardContentView.swift
